Update post_process.py
Set segmentation_mask parameter as optional. Implementation does not resize the image back to its initial size for stardist (no parameter to get the initial size).

Problem: is it necessary to resize the rescaled image back to its initial size for stardist, since we need to filter according to # pixels?